pyproject.toml
[project]
name = "desafio-tech-solutio"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.17.2",
    "bcrypt>=5.0.0",
    "flasgger>=0.9.7.1",
    "flask>=3.1.2",
    "flask-cors>=6.0.1",
    "flask-jwt-extended>=4.7.1",
    "flask-restful>=0.3.10",
    "flask-restx>=1.3.2",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "redis>=7.1.0",
    "sqlalchemy>=2.0.44",
]
